# feat(perf): bulk content write api