--- a/put.js
+++ b/put.js
@@ -1,31 +1,28 @@
 'use strict'
 
-const Promise = require('bluebird')
-
 const index = require('./lib/entry-index')
 const memo = require('./lib/memoization')
-const pipe = Promise.promisify(require('mississippi').pipe)
-const putContent = require('./lib/content/put-stream')
-const through = require('mississippi').through
+const write = require('./lib/content/write')
 const to = require('mississippi').to
 
 module.exports = putData
 function putData (cache, key, data, opts) {
   opts = opts || {}
-  const src = through()
-  let digest
-  const dest = putStream(cache, key, opts)
-  dest.on('digest', d => { digest = d })
-  const ret = pipe(src, dest).then(() => digest)
-  src.write(data, () => src.end())
-  return ret
+  return write(cache, data, opts).then(digest => {
+    return index.insert(cache, key, digest, opts).then(entry => {
+      if (opts.memoize) {
+        memo.put(cache, entry, data)
+      }
+      return digest
+    })
+  })
 }
 
 module.exports.stream = putStream
 function putStream (cache, key, opts) {
   opts = opts || {}
   let digest
-  const contentStream = putContent(cache, opts).on('digest', function (d) {
+  const contentStream = write.stream(cache, opts).on('digest', d => {
     digest = d
   })
   let memoData