Fixes: #55

### Benchmarks

```
================================================
     cacache.put()
------------------------------------------------
  1170 ops/s @ ~0.855ms/op (-83.57% ±3.42%)
  Sampled 74 in 6.00s.
================================================
     cacache.put.stream() 2048 bytes
------------------------------------------------
  1001 ops/s @ ~0.999ms/op (-87.58% ±3.46%)
  Sampled 73 in 5.88s.
================================================
     cacache.put.stream() 2048000 bytes
------------------------------------------------
  35.11 ops/s @ ~28.480ms/op (-1.96% ±0.99%)
  Sampled 351 in 34.51s.
================================================
```